rewrite pdp8.c to use state machine
I'm not happy with the previous approach, so here's a new one: first we
care about the dirty state and defer the DCA until needed.  Then we give
the instructions to be generated into a state machine that defers
emission until an appropriate point of time and remembers invariants
about the AC state, giving it the ability to perform peephole
optimisations.  The state machine is written in a coroutine kind of way,
so it looks a bit uggly, but it should do the trick.  The state machine
will be in isel() and should come in a future commit.