implement skip forwarding
When a skip instruction is followed by (CLA) IAC to copy the result of the
skip into AC and that is followed by an SNA or SZA, we discard the (CLA) IAC
and the SNA/SZA and forward the first skip instruction to the eventual
branch.

This typically saves one or two instructions in conditional branches.