turn special instructions into separate functions
I originally used special instructions because all of isel was
contained within a single function with static variables in it
being the whole state of the isel state machine.  Now that this
design has been changed to a whole translation unit, there is
no point in having special instructions where a normal function
would do the trick.